requirements.txt
python-telegram-bot==20.8  # O la versión más reciente que uses 
requests